AI_Writing_Test/requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.0.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
openai==1.51.2
psycopg2==2.9.9
